wrapper around censor_family_onsets This functions accepts a tibble with family graphs and each individual. It then censors the onset times for each individual based on the proband's end of follow-up. Returns a formatted output.

wrapper around censor_family_onsets This functions accepts a tibble with family graphs and each individual. It then censors the onset times for each individual based on the proband's end of follow-up. Returns a formatted output.

Usage

censor_family_onsets_per_family(
  family_graphs,
  tbl,
  start,
  end,
  event,
  status_col,
  aod_col,
  age_eof_col,
  fam_graph_col = "fam_graph",
  fid = "fid",
  pid = "pid",
  merge_by = pid
)

Arguments

family_graphs

tibble with fid and family graphs columns

tbl

tibble with information on each considered individual

start

column name of start of follow up, typically date of birth

end

column name of the personalised end of follow up

event

column name of the event

status_col

column name of the status (to be created)

aod_col

column name of the age of diagnosis

age_eof_col

column name of the age at the end of follow up

fam_graph_col

column name of family graphs in the 'family_graphs' object

fid

family id, typically the name of the proband that a family graph is centred on

pid

personal identifier for each individual

merge_by

column names to merge by. If different names are used for family graphs and tbl, a named vector can be specified: setNames(c("id"), c("pid")). Note id is the column name in tbl and pid is the column name in family_graphs.

Value

A tibble with family ids and updated status, age of diagnosis, and age at end of follow-up for each individual in the family based on the proband's end of follow-up.

Examples