All data preparation steps in one function.

This function is a wrapper for the following functions: get_family_graphs, censor_family_onsets_per_family, prepare_LTFHPlus_input, and fam_graph_attach_attribute. See details on each function for more information.

Usage

complete_data_preparation(
  pop_graph,
  fid = "fid",
  ndegree,
  fam_graph_col = "fam_graph",
  proband_vec,
  pheno,
  start,
  end,
  event,
  status_col,
  aod_col,
  age_eof_col,
  merge_by = setNames(c("id"), c("pid")),
  pid = "pid",
  CIP,
  CIP_merge_columns = c("age", "birth_year", "sex"),
  age_col = "age",
  CIP_cip_col = "cip",
  personal_thr = TRUE,
  use_fixed_case_thr = FALSE,
  interpolation = "xgboost",
  attached_fam_graph_col = "masked_fam_graph",
  cols_to_attach = c("lower", "upper"),
  censor_proband_thrs = T
)

Arguments

pop_graph

population graph from prepare_graph()

fid

column name of family id, column typically contains the name of the proband that a family graph is centred on

ndegree

number of steps away from proband to include

fam_graph_col

column name of family graphs in the output

proband_vec

vector of proband ids to create family graphs for

pheno

tibble with information on each considered individual

start

column name of start of follow up, typically date of birth

end

column name of the personalised end of follow up

event

column name of the event, typically age at diagnosis

status_col

column name of the status (to be created)

aod_col

column name of the age of diagnosis (to be created)

age_eof_col

column name of the age at end of follow up (to be created)

merge_by

column names to merge pheno to each individual in a family by. If different names are used for family graphs and tbl, a named vector can be specified: setNames(c("id"), c("pid")). Note id is the column name in tbl and pid is the column name in family_graphs.

pid

personal identifier for each individual in a family. Allows for multiple instances of the same individual across families. Defaults to "pid".

CIP

CIP object, must contain columns specified in CIP_merge_columns and CIP_cip_col

CIP_merge_columns

columns the CIPs are stratified by, column names must be shared with pheno or a named vector must be provided to merge_by.

age_col

column with the age in the CIP object

CIP_cip_col

column with the cumulative incidence proportions (CIP) values in the CIP object

personal_thr

Should threshold be assigned with a personalised threshold based on CIPs or population prevalence? Defaults to TRUE.

use_fixed_case_thr

Should upper and lower threshold be the same for cases? Defaults to FALSE.

interpolation

type of interpolation, defaults to "xgboost". See prepare_LTFHPlus_input for more information.

attached_fam_graph_col

column name of the updated family graphs with attached attributes.

cols_to_attach

columns to attach to the family graphs, typically lower and upper thresholds

censor_proband_thrs

should proband thresholds be censored? Used for prediction to exclude information on the proband. Defaults to TRUE.

mindist

minimum distance from proband to include in the graph (experimental, untested), defaults to 0, passed directly to make_neighborhood_graph

mode

type of distance measure in the graph (experimental, untested), defaults to "all", passed directly to make_neighborhood_graph

Value

A tibble with two columns, family ids and an updated family graph with attached attributes. If lower and upper thresholds are specified, the input is ready for estimate_liability().

Examples