Wrapper to attach attributes to family graphs

This function can attach attributes to family graphs, such as lower and upper thresholds, for each family member. This allows for personalised thresholds and other per-family specific attributes. This function wraps around assign_family_specific_thresholds to ease the process of attaching attributes to family graphs in the standard format.

Usage

fam_graph_attach_attribute(
  family_graphs,
  fam_attr,
  fam_graph_col = "fam_graph",
  attached_fam_graph_col = "masked_fam_graph",
  fid = "fid",
  pid = "pid",
  cols_to_attach = c("lower", "upper"),
  censor_proband_thrs = T
)

Arguments

family_graphs

tibble with family ids and family graphs

fam_attr

tibble with attributes for each family member

fam_graph_col

column name of family graphs in family_graphs. defailts to "fam_graph"

attached_fam_graph_col

column name of the updated family graphs with attached attributes. defaults to "masked_fam_graph".

fid

column name of family id. Typically contains the name of the proband that a family graph is centred on. defaults to "fid".

pid

personal identifier for each individual in a family. Allows for multiple instances of the same individual across families. Defaults to "pid".

cols_to_attach

columns to attach to the family graphs from fam_attr, typically lower and upper thresholds. Mixture input also requires K_i and K_pop.

censor_proband_thrs

should proband thresholds be censored? Used for prediction to exclude information on the proband. Defaults to TRUE.

Value

tibble with family ids and an updated family graph with attached attributes. If lower and upper thresholds are specified, the input is ready for estimate_liability().

Examples