Simulate under the liability threshold model (multiple phenotypes).

simulate_under_LTM_multi simulates families and thresholds under the liability threshold model for a given family structure and multiple phenotypes. Please note that it is not possible to simulate different family structures.

Usage

simulate_under_LTM_multi(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  genetic_corrmat = diag(3),
  full_corrmat = diag(3),
  h2_vec = rep(0.5, 3),
  phen_names = NULL,
  n_sim = 1000,
  pop_prev = rep(0.1, 3)
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list: - m (Mother) - f (Father) - c[0-9]*.[0-9]* (Children) - mgm (Maternal grandmother) - mgf (Maternal grandfather) - pgm (Paternal grandmother) - pgf (Paternal grandfather) - s[0-9]* (Full siblings) - mhs[0-9]* (Half-siblings - maternal side) - phs[0-9]* (Half-siblings - paternal side) - mau[0-9]* (Aunts/Uncles - maternal side) - pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

genetic_corrmat

A numeric matrix holding the genetic correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

full_corrmat

A numeric matrix holding the full correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

h2_vec

A numeric vector holding the liability-scale heritabilities for a number of phenotype. All entries must be non-negative. Note that under the liability threshold model, the heritabilities must also be at most 1. Defaults to rep(0.5,3).

phen_names

A character vector holding the phenotype names. These names will be used to create the row and column names for the covariance matrix. If it is not specified, the names will default to phenotype1, phenotype2, etc. Defaults to NULL.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A numeric vector holding the population prevalences, i.e. the overall prevalences in the population. All entries in pop_prev must be positive and smaller than 1. Defaults to rep(.1,3).

Value

If either fam_vec or n_fam is used as the argument and if it is of the required format, if genetic_corrmat and full_corrmat are two numeric and symmetric matrices satisfying that all diagonal entries are one and that all off-diagonal entries are between -1 and 1, if the liability-scale heritabilities in h2_vec are numbers satisfying \(0 \leq h^2_i\) for all \(i \in \{1,...,n_pheno\}\), n_sim is a strictly positive number, and pop_prev is a positive numeric vector such that all entries are at most one, then the output will be a list containing lists for each phenotype. The first outer list, which is named after the first phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the first phenotype. As the first outer list, the second outer list, which is named after the second phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the second phenotype. There is a list containing sim_obs for each phenotype in phen_names. The last list entry, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families and all phenotypes. Note that this tibble has the format required in estimate_liability. Finally, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM_multi()
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fid    g_phenotype1 o_phenotype1 m_phenotype1 f_phenotype1 s1_phenotype1
#>    <chr>         <dbl>        <dbl>        <dbl>        <dbl>         <dbl>
#>  1 fid_1        0.437        0.706      -1.15           0.333        -0.846
#>  2 fid_2        0.327        0.393      -0.540          1.53         -0.614
#>  3 fid_3        0.109        0.960       0.330         -0.780        -0.121
#>  4 fid_4        0.520       -0.0965     -0.485          0.826         0.402
#>  5 fid_5       -0.157       -0.127       1.15          -0.392        -1.05 
#>  6 fid_6       -0.330       -0.883      -0.581          1.59          0.488
#>  7 fid_7       -0.0953      -1.10        0.248         -1.40         -2.08 
#>  8 fid_8       -0.449       -0.573       0.00883       -1.51          0.285
#>  9 fid_9        0.309       -0.0310      0.756          0.522         0.295
#> 10 fid_10      -0.903        0.361      -1.34           0.136         0.130
#> # ℹ 990 more rows
#> # ℹ 20 more variables: mgm_phenotype1 <dbl>, mgf_phenotype1 <dbl>,
#> #   pgm_phenotype1 <dbl>, pgf_phenotype1 <dbl>, o_phenotype1_status <lgl>,
#> #   m_phenotype1_status <lgl>, f_phenotype1_status <lgl>,
#> #   s1_phenotype1_status <lgl>, mgm_phenotype1_status <lgl>,
#> #   mgf_phenotype1_status <lgl>, pgm_phenotype1_status <lgl>,
#> #   pgf_phenotype1_status <lgl>, o_phenotype1_aoo <dbl>, …
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fid    g_phenotype2 o_phenotype2 m_phenotype2 f_phenotype2 s1_phenotype2
#>    <chr>         <dbl>        <dbl>        <dbl>        <dbl>         <dbl>
#>  1 fid_1        0.816        1.24        -0.109        -0.397       0.00529
#>  2 fid_2       -0.814       -0.746        1.28         -1.88       -0.963  
#>  3 fid_3        0.683       -0.516       -0.0380        1.23        0.657  
#>  4 fid_4        1.23         1.33         0.535         0.858       1.60   
#>  5 fid_5       -0.145       -0.0316       0.641         0.776      -0.0948 
#>  6 fid_6        0.0983       0.380        0.712         1.27       -0.966  
#>  7 fid_7        0.858        0.346       -0.116        -0.340       0.318  
#>  8 fid_8        0.119        1.36         0.388         0.702       0.0952 
#>  9 fid_9        0.625        0.583       -0.529        -0.355       1.39   
#> 10 fid_10       0.784        1.40         1.54          0.436       0.784  
#> # ℹ 990 more rows
#> # ℹ 20 more variables: mgm_phenotype2 <dbl>, mgf_phenotype2 <dbl>,
#> #   pgm_phenotype2 <dbl>, pgf_phenotype2 <dbl>, o_phenotype2_status <lgl>,
#> #   m_phenotype2_status <lgl>, f_phenotype2_status <lgl>,
#> #   s1_phenotype2_status <lgl>, mgm_phenotype2_status <lgl>,
#> #   mgf_phenotype2_status <lgl>, pgm_phenotype2_status <lgl>,
#> #   pgf_phenotype2_status <lgl>, o_phenotype2_aoo <dbl>, …
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fid    g_phenotype3 o_phenotype3 m_phenotype3 f_phenotype3 s1_phenotype3
#>    <chr>         <dbl>        <dbl>        <dbl>        <dbl>         <dbl>
#>  1 fid_1      -0.0241        -0.542      -0.992       -1.48          1.38  
#>  2 fid_2      -0.00110        0.780      -0.539        0.353         0.0513
#>  3 fid_3      -0.904         -1.84       -0.187       -0.943        -0.411 
#>  4 fid_4      -1.27          -0.695      -0.995        1.14         -1.64  
#>  5 fid_5      -0.206          0.445       1.91        -1.99          1.15  
#>  6 fid_6       0.610          0.879       0.135        1.39         -1.28  
#>  7 fid_7       0.491         -0.552       1.71        -0.416         0.777 
#>  8 fid_8       1.39           2.21        0.277       -0.664        -0.344 
#>  9 fid_9       1.46          -0.444       0.0497       0.764         1.50  
#> 10 fid_10      1.31           0.770       0.416       -0.0386        0.322 
#> # ℹ 990 more rows
#> # ℹ 20 more variables: mgm_phenotype3 <dbl>, mgf_phenotype3 <dbl>,
#> #   pgm_phenotype3 <dbl>, pgf_phenotype3 <dbl>, o_phenotype3_status <lgl>,
#> #   m_phenotype3_status <lgl>, f_phenotype3_status <lgl>,
#> #   s1_phenotype3_status <lgl>, mgm_phenotype3_status <lgl>,
#> #   mgf_phenotype3_status <lgl>, pgm_phenotype3_status <lgl>,
#> #   pgf_phenotype3_status <lgl>, o_phenotype3_aoo <dbl>, …
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fid    indiv_ID role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>  <chr>    <chr>            <dbl>            <dbl>            <dbl>
#>  1 fid_1  fid_1_1  o                 -Inf             2.72          -Inf   
#>  2 fid_2  fid_2_1  o                 -Inf             2.95          -Inf   
#>  3 fid_3  fid_3_1  o                 -Inf             2.47          -Inf   
#>  4 fid_4  fid_4_1  o                 -Inf             3.31             1.33
#>  5 fid_5  fid_5_1  o                 -Inf             2.83          -Inf   
#>  6 fid_6  fid_6_1  o                 -Inf             2.95          -Inf   
#>  7 fid_7  fid_7_1  o                 -Inf             2.76          -Inf   
#>  8 fid_8  fid_8_1  o                 -Inf             3.38             1.36
#>  9 fid_9  fid_9_1  o                 -Inf             3.42          -Inf   
#> 10 fid_10 fid_10_1 o                 -Inf             3.28             1.39
#> # ℹ 7,990 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#> 

genetic_corrmat <- matrix(0.4, 3, 3)
diag(genetic_corrmat) <- 1
full_corrmat <- matrix(0.6, 3, 3)
diag(full_corrmat) <- 1

simulate_under_LTM_multi(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2,2),
c("m","mgm","mgf","s","mhs")))
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fid    g_phenotype1 o_phenotype1 m_phenotype1 mgm_phenotype1 mgf_phenotype1
#>    <chr>         <dbl>        <dbl>        <dbl>          <dbl>          <dbl>
#>  1 fid_1        1.29         1.50         0.893         -1.41            1.21 
#>  2 fid_2       -0.302        0.0844      -0.947          0.260           1.05 
#>  3 fid_3        0.850        1.12         0.189          0.414          -1.62 
#>  4 fid_4        1.43         0.798        2.73           1.82           -0.518
#>  5 fid_5        0.426        1.28        -0.500         -0.397           2.15 
#>  6 fid_6        0.0930       0.964       -0.936          0.407          -0.508
#>  7 fid_7        0.568       -0.702       -0.361         -0.745          -2.01 
#>  8 fid_8        1.40         1.30         0.816         -0.0860         -0.364
#>  9 fid_9       -0.243       -1.50         0.370         -0.977          -0.766
#> 10 fid_10      -0.821       -0.828       -0.0802         0.677          -0.464
#> # ℹ 990 more rows
#> # ℹ 20 more variables: s1_phenotype1 <dbl>, s2_phenotype1 <dbl>,
#> #   mhs1_phenotype1 <dbl>, mhs2_phenotype1 <dbl>, o_phenotype1_status <lgl>,
#> #   m_phenotype1_status <lgl>, mgm_phenotype1_status <lgl>,
#> #   mgf_phenotype1_status <lgl>, s1_phenotype1_status <lgl>,
#> #   s2_phenotype1_status <lgl>, mhs1_phenotype1_status <lgl>,
#> #   mhs2_phenotype1_status <lgl>, o_phenotype1_aoo <dbl>, …
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fid    g_phenotype2 o_phenotype2 m_phenotype2 mgm_phenotype2 mgf_phenotype2
#>    <chr>         <dbl>        <dbl>        <dbl>          <dbl>          <dbl>
#>  1 fid_1       -0.258       -0.334        -0.243          0.616         0.713 
#>  2 fid_2        0.521       -0.169         0.533          1.59          0.585 
#>  3 fid_3        1.45        -0.221        -0.756          0.993        -0.535 
#>  4 fid_4        0.773        1.13          1.26          -0.198        -1.32  
#>  5 fid_5        0.516        0.893        -1.97          -1.21         -0.264 
#>  6 fid_6       -0.308       -0.161        -1.40           1.00         -0.0462
#>  7 fid_7       -0.222        0.271         0.123          0.390        -1.27  
#>  8 fid_8        0.0857      -0.0301       -1.46           0.716        -0.751 
#>  9 fid_9       -0.166       -0.266         0.137          0.543        -0.623 
#> 10 fid_10       0.367        0.295         0.499          0.513         0.509 
#> # ℹ 990 more rows
#> # ℹ 20 more variables: s1_phenotype2 <dbl>, s2_phenotype2 <dbl>,
#> #   mhs1_phenotype2 <dbl>, mhs2_phenotype2 <dbl>, o_phenotype2_status <lgl>,
#> #   m_phenotype2_status <lgl>, mgm_phenotype2_status <lgl>,
#> #   mgf_phenotype2_status <lgl>, s1_phenotype2_status <lgl>,
#> #   s2_phenotype2_status <lgl>, mhs1_phenotype2_status <lgl>,
#> #   mhs2_phenotype2_status <lgl>, o_phenotype2_aoo <dbl>, …
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fid    g_phenotype3 o_phenotype3 m_phenotype3 mgm_phenotype3 mgf_phenotype3
#>    <chr>         <dbl>        <dbl>        <dbl>          <dbl>          <dbl>
#>  1 fid_1       -1.03        -0.424        -0.835          0.435        -0.0704
#>  2 fid_2        0.889       -0.316        -1.66          -1.16          0.115 
#>  3 fid_3        0.167       -0.913        -1.16           0.220         0.295 
#>  4 fid_4       -0.0735       1.33          1.02           1.31          0.612 
#>  5 fid_5        0.609        1.74          1.60          -1.43         -0.582 
#>  6 fid_6        0.568        1.02          1.13           0.848         1.07  
#>  7 fid_7       -0.397        0.0100        0.305          0.505         0.843 
#>  8 fid_8       -1.05        -1.70         -0.752         -0.618        -1.16  
#>  9 fid_9       -0.300        0.975         0.909         -0.115         0.553 
#> 10 fid_10      -0.867       -0.480        -0.736         -0.424         0.284 
#> # ℹ 990 more rows
#> # ℹ 20 more variables: s1_phenotype3 <dbl>, s2_phenotype3 <dbl>,
#> #   mhs1_phenotype3 <dbl>, mhs2_phenotype3 <dbl>, o_phenotype3_status <lgl>,
#> #   m_phenotype3_status <lgl>, mgm_phenotype3_status <lgl>,
#> #   mgf_phenotype3_status <lgl>, s1_phenotype3_status <lgl>,
#> #   s2_phenotype3_status <lgl>, mhs1_phenotype3_status <lgl>,
#> #   mhs2_phenotype3_status <lgl>, o_phenotype3_aoo <dbl>, …
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fid    indiv_ID role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>  <chr>    <chr>            <dbl>            <dbl>            <dbl>
#>  1 fid_1  fid_1_1  o                 1.49             1.49             -Inf
#>  2 fid_2  fid_2_1  o              -Inf                3.45             -Inf
#>  3 fid_3  fid_3_1  o              -Inf                2.72             -Inf
#>  4 fid_4  fid_4_1  o              -Inf                2.68             -Inf
#>  5 fid_5  fid_5_1  o                 1.28             1.28             -Inf
#>  6 fid_6  fid_6_1  o              -Inf                2.91             -Inf
#>  7 fid_7  fid_7_1  o              -Inf                2.79             -Inf
#>  8 fid_8  fid_8_1  o                 1.30             1.30             -Inf
#>  9 fid_9  fid_9_1  o              -Inf                3.24             -Inf
#> 10 fid_10 fid_10_1 o              -Inf                3.31             -Inf
#> # ℹ 7,990 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#> 

simulate_under_LTM_multi(fam_vec = c("m","f","s1"), add_ind = FALSE,
genetic_corrmat = genetic_corrmat, full_corrmat = full_corrmat, n_sim = 100)
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 100 × 10
#>    fid    m_phenotype1 f_phenotype1 s1_phenotype1 m_phenotype1_status
#>    <chr>         <dbl>        <dbl>         <dbl> <lgl>              
#>  1 fid_1        -0.329       -0.182         1.72  FALSE              
#>  2 fid_2        -1.29        -1.07          0.427 FALSE              
#>  3 fid_3        -0.525       -1.34          0.144 FALSE              
#>  4 fid_4        -0.187       -0.441         0.687 FALSE              
#>  5 fid_5        -1.40         1.55         -1.52  FALSE              
#>  6 fid_6        -0.804       -0.803        -1.52  FALSE              
#>  7 fid_7         0.768        1.91          0.345 FALSE              
#>  8 fid_8        -1.26         0.388        -0.286 FALSE              
#>  9 fid_9         0.772       -0.402         0.199 FALSE              
#> 10 fid_10       -0.239       -1.05         -1.37  FALSE              
#> # ℹ 90 more rows
#> # ℹ 5 more variables: f_phenotype1_status <lgl>, s1_phenotype1_status <lgl>,
#> #   m_phenotype1_aoo <dbl>, f_phenotype1_aoo <dbl>, s1_phenotype1_aoo <dbl>
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 100 × 10
#>    fid    m_phenotype2 f_phenotype2 s1_phenotype2 m_phenotype2_status
#>    <chr>         <dbl>        <dbl>         <dbl> <lgl>              
#>  1 fid_1     -0.871           0.705         1.17  FALSE              
#>  2 fid_2     -0.706          -2.17         -0.990 FALSE              
#>  3 fid_3      0.102          -1.28         -0.330 FALSE              
#>  4 fid_4      0.919           0.285        -1.43  FALSE              
#>  5 fid_5     -0.949           0.590        -1.70  FALSE              
#>  6 fid_6      0.610          -0.603        -1.71  FALSE              
#>  7 fid_7      0.396           2.31          0.206 FALSE              
#>  8 fid_8     -0.000633       -0.111        -0.272 FALSE              
#>  9 fid_9     -0.176           0.169        -0.432 FALSE              
#> 10 fid_10    -1.82           -0.978        -1.68  FALSE              
#> # ℹ 90 more rows
#> # ℹ 5 more variables: f_phenotype2_status <lgl>, s1_phenotype2_status <lgl>,
#> #   m_phenotype2_aoo <dbl>, f_phenotype2_aoo <dbl>, s1_phenotype2_aoo <dbl>
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 100 × 10
#>    fid    m_phenotype3 f_phenotype3 s1_phenotype3 m_phenotype3_status
#>    <chr>         <dbl>        <dbl>         <dbl> <lgl>              
#>  1 fid_1         1.01        0.532         2.20   FALSE              
#>  2 fid_2        -1.08       -0.645         0.187  FALSE              
#>  3 fid_3         0.565      -2.77         -1.19   FALSE              
#>  4 fid_4        -0.959       0.303        -0.0441 FALSE              
#>  5 fid_5         1.52        0.190        -0.971  TRUE               
#>  6 fid_6        -0.955      -0.788        -2.12   FALSE              
#>  7 fid_7         0.553       0.529        -0.468  FALSE              
#>  8 fid_8         0.327      -0.0354       -0.563  FALSE              
#>  9 fid_9         0.874       1.13          1.17   FALSE              
#> 10 fid_10       -0.393       0.387        -0.506  FALSE              
#> # ℹ 90 more rows
#> # ℹ 5 more variables: f_phenotype3_status <lgl>, s1_phenotype3_status <lgl>,
#> #   m_phenotype3_aoo <dbl>, f_phenotype3_aoo <dbl>, s1_phenotype3_aoo <dbl>
#> 
#> 
#> $thresholds
#> # A tibble: 300 × 9
#>    fid    indiv_ID role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>  <chr>    <chr>            <dbl>            <dbl>            <dbl>
#>  1 fid_1  fid_1_1  m                 -Inf             1.64             -Inf
#>  2 fid_2  fid_2_1  m                 -Inf             1.59             -Inf
#>  3 fid_3  fid_3_1  m                 -Inf             2.18             -Inf
#>  4 fid_4  fid_4_1  m                 -Inf             2.09             -Inf
#>  5 fid_5  fid_5_1  m                 -Inf             1.71             -Inf
#>  6 fid_6  fid_6_1  m                 -Inf             1.64             -Inf
#>  7 fid_7  fid_7_1  m                 -Inf             2.22             -Inf
#>  8 fid_8  fid_8_1  m                 -Inf             1.74             -Inf
#>  9 fid_9  fid_9_1  m                 -Inf             2.30             -Inf
#> 10 fid_10 fid_10_1 m                 -Inf             2.01             -Inf
#> # ℹ 290 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#> 

simulate_under_LTM_multi(fam_vec = c(), n_fam = NULL, add_ind = TRUE, n_sim = 150)
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 150 × 5
#>    fid    g_phenotype1 o_phenotype1 o_phenotype1_status o_phenotype1_aoo
#>    <chr>         <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fid_1       -1.18        -1.37   FALSE                             22
#>  2 fid_2        0.0328      -0.0414 FALSE                             28
#>  3 fid_3        0.614        1.15   FALSE                             37
#>  4 fid_4        0.288        1.29   TRUE                              92
#>  5 fid_5        0.539        0.938  FALSE                             32
#>  6 fid_6        0.170        0.388  FALSE                             21
#>  7 fid_7        0.376        1.14   FALSE                             34
#>  8 fid_8       -0.517       -0.832  FALSE                             31
#>  9 fid_9        0.614        0.206  FALSE                             26
#> 10 fid_10      -0.369        0.145  FALSE                             34
#> # ℹ 140 more rows
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 150 × 5
#>    fid    g_phenotype2 o_phenotype2 o_phenotype2_status o_phenotype2_aoo
#>    <chr>         <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fid_1       -0.0286       -0.964 FALSE                             22
#>  2 fid_2        0.624         0.833 FALSE                             28
#>  3 fid_3        1.52          2.70  TRUE                              33
#>  4 fid_4        0.114         1.18  FALSE                             20
#>  5 fid_5        0.562        -0.195 FALSE                             32
#>  6 fid_6       -1.47         -0.885 FALSE                             21
#>  7 fid_7       -0.577        -1.37  FALSE                             34
#>  8 fid_8       -0.215        -1.29  FALSE                             31
#>  9 fid_9        0.247         0.686 FALSE                             26
#> 10 fid_10       0.283         0.673 FALSE                             34
#> # ℹ 140 more rows
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 150 × 5
#>    fid    g_phenotype3 o_phenotype3 o_phenotype3_status o_phenotype3_aoo
#>    <chr>         <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fid_1       -0.487       -0.890  FALSE                             22
#>  2 fid_2        0.909        0.943  FALSE                             28
#>  3 fid_3        0.0173       0.873  FALSE                             37
#>  4 fid_4        0.150        0.577  FALSE                             20
#>  5 fid_5       -0.439       -0.484  FALSE                             32
#>  6 fid_6        0.436        0.784  FALSE                             21
#>  7 fid_7        0.104        0.0635 FALSE                             34
#>  8 fid_8       -0.114       -0.621  FALSE                             31
#>  9 fid_9       -0.396       -1.18   FALSE                             26
#> 10 fid_10      -1.01        -0.906  FALSE                             34
#> # ℹ 140 more rows
#> 
#> 
#> $thresholds
#> # A tibble: 150 × 9
#>    fid    indiv_ID role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>  <chr>    <chr>            <dbl>            <dbl>            <dbl>
#>  1 fid_1  fid_1_1  o              -Inf                3.14          -Inf   
#>  2 fid_2  fid_2_1  o              -Inf                2.91          -Inf   
#>  3 fid_3  fid_3_1  o              -Inf                2.55             2.72
#>  4 fid_4  fid_4_1  o                 1.29             1.29          -Inf   
#>  5 fid_5  fid_5_1  o              -Inf                2.76          -Inf   
#>  6 fid_6  fid_6_1  o              -Inf                3.17          -Inf   
#>  7 fid_7  fid_7_1  o              -Inf                2.68          -Inf   
#>  8 fid_8  fid_8_1  o              -Inf                2.79          -Inf   
#>  9 fid_9  fid_9_1  o              -Inf                2.99          -Inf   
#> 10 fid_10 fid_10_1 o              -Inf                2.68          -Inf   
#> # ℹ 140 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#>